Restructured the filetree of the /moduels folder
