--- phrasechecker.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <script type="text/javascript" src="dist/js/app.js"></script>
-    <script type="text/babel" src="src/js/phraseChecker.js"></script>
-    <link rel="stylesheet" href="styles/bootstrap.min.css">
-    <link rel="stylesheet" href="styles/style.css">
-  <head>
-  <body>
-    <div id="module"></div>
-  </body>
-
-</html>
