+++ src/js/components/modules/phrase_check_module/ReportView.js
@@ -1,0 +1,21 @@
+var React = require('react');
+
+var ReportObject = {};
+
+var PhraseReport = React.createClass({
+  getInitialState: function(){
+    return {
+      info: reportObject
+    }
+  },
+  render: function(){
+    for(let check in reportObject){
+      
+    }
+    return (
+
+    );
+  }
+});
+
+module.exports = PhraseReport;
