+++ src/js/components/modules/phrase_check_module/subcomponents/ConfirmDisplay.js
@@ -1,0 +1,15 @@
+var React = require('react');
+
+var ConfirmDisplay = React.createClass({
+  render: function(){
+    return (
+      <form>
+        <label>{this.props.toCheck}</label>
+        <label>{this.props.note}</label>
+      </form>
+    );
+  }
+});
+
+
+module.exports = ConfirmDisplay;
