+++ src/js/components/modules/phrase_check_module/subcomponents/Loader.js
@@ -1,0 +1,16 @@
+var React = require('react');
+var RB = require('react-bootstrap');
+var {ProgressBar} = RB;
+
+var Loader = React.createClass({
+  render: function(){
+    return (
+      <ProgressBar
+        now={this.props.progress}
+        style={{verticaAlign: 'middle'}}
+      />
+    );
+  }
+});
+
+module.exports = Loader;
